ai-scaling-achitechture/worker-service/requirements.txt
google-cloud-pubsub==2.18.0
google-cloud-storage==2.7.0
requests==2.28.1